Added insert and search methods

+++ src/symbol_table.h
@@ -18,8 +18,8 @@ typedef struct s_Symbol_Table
 
 void st_init(Symbol_Table *st);
 void st_free(Symbol_Table *st);
-void st_insert(Symbol_Table *st, char *label, int memory_addr);
-int st_search(Symbol_Table *st, char *label);
+void st_insert(Symbol_Table *st, char *label, uint8_t memory_addr);
+uint8_t st_search(Symbol_Table *st, char *label);
 void st_remove(Symbol_Table *st, char *label);
 
 #endif /* SYMBOL_TABLE_H_ */
